+++ R/airdas_process.R
@@ -3,9 +3,9 @@
 #' Process AirDAS data (the output of \code{\link{airdas_read}}), 
 #'   including extracting state and condition information for each AirDAS event
 #'
-#' @param x either an object of class \code{airdas_dfr} (the output of \code{\link{airdas_read}}), 
-#'   or a character (filepath) which is first passed to 
-#'   the \code{file} argument of \code{\link{airdas_read}}
+#' @param x an object of class \code{airdas_dfr} object, 
+#'   an object that can be coerced to class \code{airdas_dfr},
+#'   or a character (filepath) which is first passed to \code{\link{airdas_read}}
 #' @param ... passed to \code{\link{airdas_read}} if \code{x} is a character.
 #'   Otherwise ignored
 #' @param days.gap.part numeric of length 1;